Loggerの初期化を別クラスに移動
ついでに、ログのアーカイブを起動毎に行うようにした
